requirements.txt
Django==4.1.3
djangorestframework==3.14.0
psycopg2==2.9.5
environs~=9.5.0
jwt~=1.3.1
djangorestframework-simplejwt~=5.3.1
Pillow~=10.1.0